feat: refactor types
