+++ packages/schema/src/accounts/author/getAll.ts
@@ -1,0 +1,1 @@
+export const route = 'author/get-all';
