+++ packages/schema/src/accounts/author/getTotal.ts
@@ -1,0 +1,1 @@
+export const route = 'author/get-total';
