+++ packages/schema/src/accounts/author/index.ts
@@ -1,0 +1,2 @@
+export * as getAll from './getAll';
+export * as getTotal from './getTotal';
