+++ packages/schema/src/accounts/index.ts
@@ -1,11 +1,2 @@
-import {routes as authorRoutes} from './author';
-import {routes as userRoutes} from './user';
-
-export namespace Accounts {
-    export namespace Author {
-        export const routes = authorRoutes;
-    }
-    export namespace User {
-        export const routes = userRoutes;
-    }
-}
+export * as author from 'author';
+export * as user from 'user';
