--- packages/schema/src/accounts/user.ts
@@ -1,6 +1,0 @@
-export const routes = {
-    login: 'users/login',
-    register: 'users/register',
-    getAll: 'users/get-all',
-    getTotal:'users/get-total',
-}
