+++ packages/schema/src/accounts/user/getAll.ts
@@ -1,0 +1,1 @@
+export const route = 'user/get-all';
