+++ packages/schema/src/accounts/user/getTotal.ts
@@ -1,0 +1,1 @@
+export const route = 'user/get-total';
