+++ packages/schema/src/accounts/user/index.ts
@@ -1,0 +1,4 @@
+export * as getAll from './getAll';
+export * as getTotal from './getTotal';
+export * as login from './login';
+export * as register from './register';
