+++ packages/schema/src/accounts/user/login.ts
@@ -1,0 +1,1 @@
+export const route = 'user/login';
