+++ packages/schema/src/accounts/user/register.ts
@@ -1,0 +1,1 @@
+export const route = 'user/register';
