--- packages/schema/src/documents/document.ts
@@ -1,5 +1,0 @@
-export const routes = {
-    getAll: 'documents/get-all',
-    getById: 'documents/get-by-id',
-    getTotal: 'documents/get-total',
-};
