+++ packages/schema/src/documents/document/getAll.ts
@@ -1,0 +1,1 @@
+export const route = 'document/get-all';
