+++ packages/schema/src/documents/document/getById.ts
@@ -1,0 +1,1 @@
+export const route = 'document/get-by-id';
