+++ packages/schema/src/documents/document/getTotal.ts
@@ -1,0 +1,1 @@
+export const route = 'document/get-total';
