+++ packages/schema/src/documents/document/index.ts
@@ -1,0 +1,3 @@
+export * as getAll from './getAll';
+export * as getById from './getById';
+export * as getTotal from './getTotal';
