+++ packages/schema/src/documents/index.ts
@@ -1,7 +1,1 @@
-import {routes as documentRoutes} from './document';
-
-export namespace Documents {
-    export namespace Document {
-        export const routes = documentRoutes;
-    }
-}
+export * as document from 'document';
