+++ packages/schema/src/system/migration/getAll.ts
@@ -1,0 +1,8 @@
+type Migration = {
+    id: string;
+    name: string;
+    date: Date;
+    success: boolean;
+}
+export type Response = Migration[];
+export const route = 'migrations/get-all';
