+++ packages/schema/src/system/migration/index.ts
@@ -1,0 +1,1 @@
+export * as getAll from './getAll';
