+++ packages/schema/src/system/search/index.ts
@@ -1,0 +1,3 @@
+export * as indexDocuments from './indexDocuments';
+export * as query from './query';
+export * as search from './search';
