+++ packages/schema/src/system/search/indexDocuments.ts
@@ -1,0 +1,2 @@
+export type Response = boolean;
+export const route = 'search/index-documents';
