+++ packages/schema/src/system/search/search.ts
@@ -1,0 +1,9 @@
+export namespace Search {
+
+
+    export namespace IndexDocuments {
+    }
+
+    export namespace Query {
+    }
+}
